backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
tensorflow
numpy
pandas
nltk==3.9.0
scikit-learn
python-dotenv==1.0.1
sqlalchemy==2.0.23
openai
